Optimize module dependencies
