+++ winnow/utils/utils.py
@@ -1,12 +1,17 @@
 import hashlib
+import json
 import os
 from glob import glob
 from pathlib import Path
+
 import cv2
 import numpy as np
+from joblib import load
+
 from winnow.config import Config
 from winnow.config.path import resolve_config_path
-from joblib import load
+from winnow.storage.repr_key import ReprKey
+from winnow.storage.repr_utils import path_resolver
 
 DEFAULT_DIRECTORY = os.path.join(os.path.dirname(__file__), "models")
 GRAY_ESTIMATION_MODEL = os.path.join(DEFAULT_DIRECTORY, "gb_gray_model.joblib")
@@ -175,3 +180,40 @@ def resolve_config(config_path=None, frame_sampling=None, save_frames=None):
     cond1 = save_frames is None and config.proc.save_frames
     config.proc.save_frames = (cond1 or save_frames)
     return config
+
+
+def get_config_tag(config):
+    """Get configuration tag.
+
+    Whenever configuration changes making the intermediate representation
+    incompatible the tag value will change as well.
+    """
+
+    # Configuration attributes that affect representation value
+    config_attributes = dict(
+        frame_sampling=config.proc.frame_sampling
+    )
+
+    sha256 = hashlib.sha256()
+    sha256.update(json.dumps(config_attributes).encode("utf-8"))
+    return sha256.hexdigest()[:40]
+
+
+def reprkey_resolver(config):
+    """Create a function to get intermediate storage key and tags by the file path.
+
+    Args:
+        config (winnow.config.Config): Pipeline configuration.
+    """
+
+    storepath = path_resolver(config.sources.root)
+    config_tag = get_config_tag(config)
+
+    def reprkey(path):
+        """Get intermediate representation storage key."""
+        return ReprKey(
+            path=storepath(path),
+            hash=get_hash(path),
+            tag=config_tag)
+
+    return reprkey
